rename: page 레벨로 변경
